requirements.txt

asgiref==3.6.0
Django==4.2
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==23.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.2
mysql-connector==2.2.9
pillow==8.2.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
